requirements.txt
fastmcp>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
loguru>=0.7.2
websockets>=11.0
